+++ src/utils/getTableRows.ts
@@ -7,16 +7,20 @@ import { getTestTags } from "./getTestTags";
 import { getTestAnnotations } from "./getTestAnnotations";
 import { getTestDuration } from "./getTestDuration";
 import { getTestStatusIcon } from "./getTestStatusIcon";
-import { DisplayLevel } from "../models";
+import { BlobService, DisplayLevel } from "../models";
+import { processAttachments } from "./processAttachments";
 
 export const getTableRows = async (
   tests: TestCase[],
   showAnnotations: boolean,
   showTags: boolean,
   showError: boolean,
-  displayLevel: DisplayLevel[]
+  displayLevel: DisplayLevel[],
+  showAnnotationsInColumn: boolean = false,
+  blobService?: BlobService
 ): Promise<SummaryTableRow[]> => {
   const convert = new Convert();
+  const hasBlobService = blobService && blobService.azure;
 
   const tableHeaders = [
     {
@@ -44,11 +48,25 @@ export const getTableRows = async (
     });
   }
 
+  if (showAnnotations && showAnnotationsInColumn){
+    tableHeaders.push({
+      data: "Annotations",
+      header: true,
+    });
+  }
+
   if (showError) {
     tableHeaders.push({
       data: "Error",
       header: true,
     });
+
+    if (hasBlobService) {
+      tableHeaders.push({
+        data: "Attachments",
+        header: true,
+      });
+    }
   }
 
   const tableRows: SummaryTableRow[] = [];
@@ -63,13 +81,17 @@ export const getTableRows = async (
       continue;
     }
 
-    if (showAnnotations && test.annotations) {
+    if (showAnnotations && !showAnnotationsInColumn && test.annotations) {
       let colLength = 4;
       if (showTags) {
         colLength++;
       }
       if (showError) {
         colLength++;
+
+        if (hasBlobService) {
+          colLength++;
+        }
       }
 
       const annotations = await getTestAnnotations(test);
@@ -110,12 +132,44 @@ export const getTableRows = async (
       });
     }
 
+    if(showAnnotations && showAnnotationsInColumn) {
+      const annotations = await getTestAnnotations(test);
+      if (annotations) {
+        tableRow.push({
+          data: annotations,
+          header: false,
+        });
+      }else{
+        tableRow.push({
+          data: "",
+          header: false,
+        });
+      }
+    }
+
     if (showError) {
       const error = result?.error?.message || "";
       tableRow.push({
         data: convert.toHtml(error),
         header: false,
       });
+
+      if (hasBlobService) {
+        const mediaFiles =
+          (await processAttachments(blobService, result.attachments)) || [];
+
+        tableRow.push({
+          data: (mediaFiles || [])
+            .map(
+              (
+                m
+              ) => `<p align="center"><img src="${m.url}" alt="${m.name}" width="250"></p>
+<p align="center"><b>${m.name}</b></p>`
+            )
+            .join("\n\n"),
+          header: false,
+        });
+      }
     }
 
     tableRows.push(tableRow);